Refactor types so that rpc is not imported in core
